fixed if version is nil, changed to small button in search, update version

+++ LSN-Helper.lua
@@ -2,7 +2,7 @@ script_name('LSN-Helper')
 script_description('Los Santos News Helper (LSNH) for special project MyHome RP')
 script_author('kyrtion#7310')
 script_properties('work-in-pause')
-script_version('4.0')
+script_version('3.5')
 
 require 'lib.moonloader'
 local dlstatus = require('moonloader').download_status
@@ -48,17 +48,16 @@ local adList = {}
 
 local new, str, sizeof = imgui.new, ffi.string, ffi.sizeof
 local renderWindow = new.bool(false)
-local menuWindow = new.bool(false)
-local tab = new.int(1)
+
 local adInput = new.char[256]('')
 local searchInput = new.char[256]('')
 
-local adNick, adPrice, adText = 'Awe Some[123]', '138$', 'awesome text, dude'
+--local adNick, adPrice, adText = 'Awe Some[123]', '138$', 'awesome text, dude'
+local adNick, adPrice, adText = '', '', ''
 local notAdNick = false
 local confirm = false
 local block = false
 local copying = false
-local autoFocus = false
 local hex = '0xEEDC82'
 
 local update_state = false
@@ -68,12 +67,12 @@ local lockFailed = false
 local newVersion = 'None'
 local oldVersion = 'None'
 
--- --! origin/master
--- local update_url = 'https://raw.githubusercontent.com/kyrtion/LSNHelper_mhrp/master/version_lsn.ini'
--- local update_path = getWorkingDirectory() .. '/update_lsn.ini'
--- local script_vers = tostring(thisScript().version)
--- local script_url = 'https://github.com/kyrtion/LSNHelper_mhrp/blob/master/LSN-Helper.lua?raw=true'
--- local script_path = thisScript().path
+--! origin/master
+local update_url = 'https://raw.githubusercontent.com/kyrtion/LSNHelper_mhrp/master/version_lsn.ini'
+local update_path = getWorkingDirectory() .. '/update_lsn.ini'
+local script_vers = tostring(thisScript().version)
+local script_url = 'https://github.com/kyrtion/LSNHelper_mhrp/blob/master/LSN-Helper.lua?raw=true'
+local script_path = thisScript().path
 
 -- --! origin/beta
 -- local update_url = 'https://raw.githubusercontent.com/kyrtion/LSNHelper_mhrp/beta/version_lsn.ini'
@@ -93,7 +92,7 @@ local newFrame = imgui.OnFrame(
 		local sizeX, sizeY = 700, 340
 		imgui.SetNextWindowPos(imgui.ImVec2(resX / 2, resY / 2), imgui.Cond.FirstUseEver, imgui.ImVec2(0.5, 0.5))
 		imgui.SetNextWindowSize(imgui.ImVec2(sizeX, sizeY * 1.04))
-		imgui.Begin(u8'Публикация oбъявления | LSN-Helper '..thisScript().version, nil, imgui.WindowFlags.NoMove + imgui.WindowFlags.NoResize + imgui.WindowFlags.NoCollapse + imgui.WindowFlags.AlwaysAutoResize)
+		imgui.Begin(u8'Публикация oбъявления | LSN-Helper '..thisScript().version, nil, imgui.WindowFlags.NoMove + imgui.WindowFlags.NoResize + imgui.WindowFlags.NoCollapse)
 
 		imgui.SetCursorPos(imgui.ImVec2(20, 40)); imgui.TextColoredRGB('Отправитель:'); imgui.SameLine((sizeX - 15) / 2 + 10); imgui.TextColoredRGB('Цена:')
 		imgui.SetCursorPos(imgui.ImVec2(20, 60));
@@ -127,7 +126,7 @@ local newFrame = imgui.OnFrame(
 
 		imgui.SetCursorPos(imgui.ImVec2(20, sizeY - 110))
 		imgui.PushItemWidth(sizeX - 40);
-
+		
 		if imgui.IsWindowAppearing() then imgui.SetKeyboardFocusHere(-1) end
 		imgui.PushAllowKeyboardFocus(false)
 		imgui.InputText(u8'##adInput', adInput, sizeof(adInput))
@@ -151,17 +150,17 @@ local newFrame = imgui.OnFrame(
 			imgui.OpenPopup(u8'Поиск')
 		end
 
-		if imgui.BeginPopupModal(u8'Поиск', _, imgui.WindowFlags.NoMove + imgui.WindowFlags.NoResize + imgui.WindowFlags.NoCollapse + imgui.WindowFlags.AlwaysAutoResize) then
-			local pSize = imgui.ImVec2(770, 360)
+		if imgui.BeginPopupModal(u8'Поиск', _, imgui.WindowFlags.NoMove + imgui.WindowFlags.NoResize + imgui.WindowFlags.NoCollapse) then
+			local pSize = imgui.ImVec2(770, 400)
 			imgui.SetWindowSizeVec2(pSize)
-	
+			
 			imgui.SetCursorPos(imgui.ImVec2(20, 45))
 			imgui.Text(u8'Текст:') imgui.SameLine(70) imgui.TextColoredRGB(adText)
-				
+			
 			imgui.SetCursorPos(imgui.ImVec2(20, 70)) imgui.Text(u8'Поиск:')
 			
 			imgui.SetCursorPos(imgui.ImVec2(70, 66))
-			imgui.PushItemWidth(pSize.x - 110)
+			imgui.PushItemWidth(pSize.x/2 + 294)
 
 			if imgui.IsWindowAppearing() then imgui.SetKeyboardFocusHere(-1) end
 			imgui.PushAllowKeyboardFocus(false)
@@ -169,16 +168,63 @@ local newFrame = imgui.OnFrame(
 			imgui.PopAllowKeyboardFocus()
 			imgui.PopItemWidth()
 			
-			imgui.SetCursorPos(imgui.ImVec2(20, 108))
+			imgui.SetCursorPos(imgui.ImVec2(21, 108))
 			imgui.Text(u8'Результаты:')
-	
-			imgui.SetCursorPos(imgui.ImVec2(19, 130))
-			imgui.BeginChild('ChildWindowsS', imgui.ImVec2(pSize.x/2 + 325, pSize.y/2 + 12), true)
+			
+			imgui.SetCursorPos(imgui.ImVec2(20, 130))
+			imgui.BeginChild('ChildWindowsS', imgui.ImVec2(pSize.x/2 + 344, pSize.y/2 + 12), true)
 			--imgui.Separator()
-			local nun = tonumber(#adList)
+			local size = imgui.GetWindowSize()
+            local ColumnsSize = {
+                [1] = 59,
+                [2] = 90,
+                [3] = 90,
+                [4] = size.x - 75 - 90 - 40 - 30,
+                [5] = 24
+            }
+
 			for i=1, #adList do
 				if string.len(str(searchInput)) ~= 0 then
 					if string.find(u8(adList[i]), str(searchInput), 1, true) then
+						imgui.Columns(2)
+						if imgui.Selectable(u8'  >  ##'..tonumber(i), true, _, imgui.ImVec2(20, 15)) then
+							imgui.StrCopy(adInput, u8(adList[i]))
+							imgui.CloseCurrentPopup()
+							searchInput = new.char[256]('')
+						end
+						imgui.SameLine(32)
+						if imgui.Selectable(u8' RB ##'..tonumber(i), true, _, imgui.ImVec2(22, 15)) then
+							sampSendChat('/rb >> '..adList[i])
+						end
+						imgui.SetColumnWidth(0, ColumnsSize[1])
+						imgui.NextColumn()
+						imgui.Text(u8(adList[i]))
+						imgui.NextColumn()
+						imgui.Separator()
+					end
+				else
+					imgui.Columns(2)
+					if imgui.Selectable(u8'  >  ##'..tonumber(i), true, _, imgui.ImVec2(20, 15)) then
+						imgui.StrCopy(adInput, u8(adList[i]))
+						imgui.CloseCurrentPopup()
+						searchInput = new.char[256]('')
+					end
+					imgui.SameLine(32)
+					if imgui.Selectable(u8' RB ##'..tonumber(i), true, _, imgui.ImVec2(22, 15)) then
+						sampSendChat('/rb >> '..adList[i])
+					end
+					imgui.SetColumnWidth(0, ColumnsSize[1])
+					imgui.NextColumn()
+					imgui.Text(u8(adList[i]))
+					imgui.NextColumn()
+					imgui.Separator()
+				end
+			end
+			-- будет второй вариант для результаты поисков
+			--[[for i=1, #adList do
+				if string.len(str(searchInput)) ~= 0 then
+					if string.find(u8(adList[i]), str(searchInput), 1, true) then
+						
 						if imgui.Button('>##'..tostring(i), imgui.ImVec2(22, 24)) then
 							imgui.StrCopy(adInput, u8(adList[i]))
 							imgui.CloseCurrentPopup()
@@ -191,8 +237,10 @@ local newFrame = imgui.OnFrame(
 						imgui.SameLine()
 						imgui.Text(u8(adList[i]))
 						if i ~= #adList then imgui.Separator() end
+
 					end
 				else
+
 					if imgui.Button('>##'..tostring(i), imgui.ImVec2(22, 24)) then
 						imgui.StrCopy(adInput, u8(adList[i]))
 						imgui.CloseCurrentPopup()
@@ -205,14 +253,15 @@ local newFrame = imgui.OnFrame(
 					imgui.SameLine()
 					imgui.Text(u8(adList[i]))
 					if i ~= #adList then imgui.Separator() end
+
 				end
-			end
+			end]]
 
 			imgui.EndChild()
 
 
-			imgui.SetCursorPos(imgui.ImVec2(5, pSize.y - 20))
-			if imgui.Button(u8'Закрыть', imgui.ImVec2(pSize.x - 30, 25)) then
+			imgui.SetCursorPos(imgui.ImVec2(20, pSize.y - 42))
+			if imgui.Button(u8'Закрыть', imgui.ImVec2(pSize.x/2 + 344, 25)) then
 				searchInput = new.char[256]('')
 				imgui.CloseCurrentPopup()
 			end
@@ -232,7 +281,7 @@ local newFrame = imgui.OnFrame(
 				send('В тексте пусто, зачем отправлять?', -1)
 
 			elseif Char and Char ~= ('$' or ',' or '/' or '>' or '<' or '-' or '=' or '+' or '_' or "'" or '"') then
-				sampSendDialogResponse(1536,1,0,(u8:decode(str(adInput))))
+				sampSendDialogResponse(1536, 1, 0, (u8:decode(str(adInput))))
 				renderWindow[0] = false
 				confirm = true
 
@@ -257,7 +306,7 @@ local newFrame = imgui.OnFrame(
 			if (u8:decode(str(adInput))) == (nil or '') then
 				send('Вы не указали причину в поле', -1)
 			else
-				sampSendDialogResponse(1536,0,0,(u8:decode(str(adInput))))
+				sampSendDialogResponse(1536, 0, 0, (u8:decode(str(adInput))))
 				
 				renderWindow[0] = false
 				copying = false
@@ -277,44 +326,6 @@ local newFrame = imgui.OnFrame(
 	end
 )
 
-local menuFrame = imgui.OnFrame(
-	function() return menuWindow[0] end,
-	function(player)
-		local resX, resY = getScreenResolution()
-        local sizeX, sizeY = 600, 400
-        imgui.SetNextWindowPos(imgui.ImVec2(resX / 2, resY / 2), imgui.Cond.FirstUseEver, imgui.ImVec2(0.5, 0.5))
-        imgui.SetNextWindowSize(imgui.ImVec2(sizeX, sizeY), imgui.Cond.FirstUseEver)
-        imgui.BeginCustomTitle(u8'LSN Helper - Версия: '..tostring(thisScript().version), 30, menuWindow, imgui.WindowFlags.NoMove + imgui.WindowFlags.NoResize + imgui.WindowFlags.NoCollapse)
-        
-        imgui.SetCursorPos(imgui.ImVec2(5, 35))
-        imgui.CustomMenu({u8'Главный', u8'Настройка', u8'Объявление', u8'Эфир'}, tab, imgui.ImVec2(75, 30), _, true)
-        imgui.SetCursorPos(imgui.ImVec2(5 + 75 + 5 + 10, 35))
-
-
-        local childSize = sizeX - 100
-        imgui.BeginChild('s', imgui.ImVec2(childSize, sizeY - 40), true)
-
-        if tab[0] == 1 then
-			imgui.Text(u8'Первый раздел')
-        elseif tab[0] == 2 then
-			imgui.Text(u8'Второй раздел')
-			imgui.Text(u8'Второй раздел')
-        elseif tab[0] == 3 then
-			imgui.Text(u8'Третий раздел')
-			imgui.Text(u8'Третий раздел')
-			imgui.Text(u8'Третий раздел')
-        end
-
-        imgui.EndChild()
-        -- if tab[0] == 1 then
-            -- imgui.Text(u8'Первый исключительный раздел')
-			-- здесь кнопки для первого раздела
-        --end
-
-        imgui.End()
-	end
-)
-
 function main()
 	while not isSampAvailable() do wait(0) end
 
@@ -327,20 +338,15 @@ function main()
 	send('Скрипт успешно загружено. Версия: '..thisScript().version)
 	print(); print('Script LSN-Helper '..thisScript().version..' loaded - Discord: kyrtion#7310')
 
-	--! debug window (dont use)
-	sampRegisterChatCommand('ef', function()
-		renderWindow[0] = not renderWindow[0]
-		imgui.StrCopy(adInput, u8(adText))
-		autoFocus = true
-	end)
-
-	sampRegisterChatCommand('mf', function()
-		menuWindow[0] = not menuWindow[0]
-	end)
+	-- --! debug window (dont use)
+	-- sampRegisterChatCommand('ef', function()
+	-- 	renderWindow[0] = not renderWindow[0]
+	-- 	imgui.StrCopy(adInput, u8(adText))
+	-- end)
 
 	sampRegisterChatCommand('verify', function()
 		if lockVerify then
-			if renderWindow[0] and sampIsDialogActive() then				
+			if renderWindow[0] or sampIsDialogActive() then				
 				send('Закройте диалог и снова вводите /verify')
 			else
 				checkVerify = true
@@ -353,15 +359,18 @@ function main()
 	downloadUrlToFile(update_url, update_path, function(id, status)
 		if status == dlstatus.STATUS_ENDDOWNLOADDATA then
 			updateIni = inicfg.load(nil, update_path)
-			newVersion = tostring(updateIni.info.version):gsub('"', '')
-			oldVersion = tostring(thisScript().version)
-			--sampAddChatMessage(newVersion..' -> '..oldVersion, -1)
-			if newVersion ~= oldVersion then
-				send('Есть обновление! Версия: '..newVersion..'. Чтобы обновить вводите /verify', -1)
-				update_state = true
-				lockVerify = true
+			if updateIni ~= nil or updateIni.info.version ~= nil then
+				newVersion = tostring(updateIni.info.version):gsub('"', '')
+				oldVersion = tostring(thisScript().version)
+				if newVersion ~= oldVersion then
+					send('Есть обновление! Версия: '..newVersion..'. Чтобы обновить вводите /verify', -1)
+					update_state = true
+					lockVerify = true
+				end
+				os.remove(update_path)
+			else
+				send('Невозможно проверить наличие обновление. Нажмите CTRL + R чтобы перезагрузить')
 			end
-			os.remove(update_path)
 		end
 	end)
 	
@@ -399,7 +408,6 @@ function sampev.onShowDialog(id, style, title, button1, button2, text)
 		adText = ( text:match('%{ffffff%}Текст%:%{7FFF00%} (.*)%{ffffff%}') ):gsub("\n", "")
 		adPrice = ( text:match('%{ffffff%}Цена%:%{7FFF00%} (.*)%{FFFFFF%}') ):gsub("\n", "")
 		renderWindow[0] = true
-		autoFocus = true
 		if editList[adText] == nil then
 			imgui.StrCopy(adInput, u8(adText))
 			copying = false
@@ -410,7 +418,7 @@ function sampev.onShowDialog(id, style, title, button1, button2, text)
 		return false
 
 	elseif id == 1537 and title == '{6333FF}Публикация объявления: {ffffff}Подтверждение' and confirm then
-		sampSendDialogResponse(1537,1,0,0)
+		sampSendDialogResponse(1537, 1, 0, 0)
 		confirm = false
 		return false
 	end
@@ -422,6 +430,7 @@ function sampev.onSendDialogResponse(dialogId, button, listboxId, input)
 			notAdNick = true
 		else
 			local fr = ''; fr, adNick = input:match('(%d+)%. (.*)')
+			notAdNick = false
 		end
 	end
 end
@@ -501,56 +510,6 @@ function sampev.onServerMessage(color, text)
 	end
 end
 
---==[ IMGUI FUNCS ]==--
--- labels - Array - названия элементов меню
--- selected - imgui.ImInt() - выбранный пункт меню
--- size - imgui.ImVec2() - размер элементов
--- speed - float - скорость анимации выбора элемента (необязательно, по стандарту - 0.2)
--- centering - bool - центрирование текста в элементе (необязательно, по стандарту - false)
-function imgui.CustomMenu(labels, selected, size, speed, centering)
-    local bool = false
-    speed = speed and speed or 0.2
-    local radius = size.y * 0.50
-    local draw_list = imgui.GetWindowDrawList()
-    if LastActiveTime == nil then LastActiveTime = {} end
-    if LastActive == nil then LastActive = {} end
-    local function ImSaturate(f)
-        return f < 0.0 and 0.0 or (f > 1.0 and 1.0 or f)
-    end
-    for i, v in ipairs(labels) do
-        local c = imgui.GetCursorPos()
-        local p = imgui.GetCursorScreenPos()
-        if imgui.InvisibleButton(v..'##'..i, size) then
-            selected[0] = i
-            LastActiveTime[v] = os.clock()
-            LastActive[v] = true
-            bool = true
-        end
-        imgui.SetCursorPos(c)
-        local t = selected[0] == i and 1.0 or 0.0
-        if LastActive[v] then
-            local time = os.clock() - LastActiveTime[v]
-            if time <= 0.3 then
-                local t_anim = ImSaturate(time / speed)
-                t = selected[0] == i and t_anim or 1.0 - t_anim
-            else
-                LastActive[v] = false
-            end
-        end
-        local col_bg = imgui.GetColorU32Vec4(selected[0] == i and imgui.GetStyle().Colors[imgui.Col.ButtonActive] or imgui.ImVec4(0,0,0,0))
-        local col_box = imgui.GetColorU32Vec4(selected[0] == i and imgui.GetStyle().Colors[imgui.Col.Button] or imgui.ImVec4(0,0,0,0))
-        local col_hovered = imgui.GetStyle().Colors[imgui.Col.ButtonHovered]
-        local col_hovered = imgui.GetColorU32Vec4(imgui.ImVec4(col_hovered.x, col_hovered.y, col_hovered.z, (imgui.IsItemHovered() and 0.2 or 0)))
-        draw_list:AddRectFilled(imgui.ImVec2(p.x-size.x/6, p.y), imgui.ImVec2(p.x + (radius * 0.65) + t * size.x, p.y + size.y), col_bg, 10.0)
-        draw_list:AddRectFilled(imgui.ImVec2(p.x-size.x/6, p.y), imgui.ImVec2(p.x + (radius * 0.65) + size.x, p.y + size.y), col_hovered, 10.0)
-        draw_list:AddRectFilled(imgui.ImVec2(p.x, p.y), imgui.ImVec2(p.x+5, p.y + size.y), col_box)
-        imgui.SetCursorPos(imgui.ImVec2(c.x+(centering and (size.x-imgui.CalcTextSize(v).x)/2 or 15), c.y+(size.y-imgui.CalcTextSize(v).y)/2))
-        imgui.Text(v)
-        imgui.SetCursorPos(imgui.ImVec2(c.x, c.y+size.y))
-    end
-    return bool
-end
-
 function imgui.TextColoredRGB(text)
 	local style = imgui.GetStyle()
 	local colors = style.Colors
@@ -603,23 +562,6 @@ function imgui.TextColoredRGB(text)
 	render_text(text)
 end
 
-function imgui.BeginCustomTitle(title, titleSizeY, var, flags)
-    imgui.PushStyleVarVec2(imgui.StyleVar.WindowPadding, imgui.ImVec2(0, 0))
-    imgui.PushStyleVarFloat(imgui.StyleVar.WindowBorderSize, 0)
-    imgui.Begin(title, var, imgui.WindowFlags.NoTitleBar + (flags or 0))
-    imgui.SetCursorPos(imgui.ImVec2(0, 0))
-    local p = imgui.GetCursorScreenPos()
-    imgui.GetWindowDrawList():AddRectFilled(p, imgui.ImVec2(p.x + imgui.GetWindowSize().x, p.y + titleSizeY), imgui.GetColorU32Vec4(imgui.GetStyle().Colors[imgui.Col.TitleBgActive]), imgui.GetStyle().WindowRounding, 1 + 2)
-    imgui.SetCursorPos(imgui.ImVec2(imgui.GetWindowSize().x / 2 - imgui.CalcTextSize(title).x / 2, titleSizeY / 2 - imgui.CalcTextSize(title).y / 2))
-    imgui.Text(title)
-    imgui.SetCursorPos(imgui.ImVec2(imgui.GetWindowSize().x - (titleSizeY - 10) - 5, 5))
-    imgui.PushStyleVarFloat(imgui.StyleVar.FrameRounding, imgui.GetStyle().WindowRounding)
-    if imgui.Button('X##CLOSEBUTTON.WINDOW.'..title, imgui.ImVec2(titleSizeY - 10, titleSizeY - 10)) then var[0] = false end
-    imgui.SetCursorPos(imgui.ImVec2(5, titleSizeY + 5))
-    imgui.PopStyleVar(3)
-    imgui.PushStyleVarVec2(imgui.StyleVar.WindowPadding, imgui.ImVec2(5, 5))
-end
-
 function imgui.DarkTheme()
 	imgui.SwitchContext()
 	--==[ STYLE ]==--
@@ -651,7 +593,7 @@ function imgui.DarkTheme()
 	--==[ ALIGN ]==--
 	imgui.GetStyle().WindowTitleAlign = imgui.ImVec2(0.5, 0.5)
 	imgui.GetStyle().ButtonTextAlign = imgui.ImVec2(0.5, 0.5)
-	imgui.GetStyle().SelectableTextAlign = imgui.ImVec2(0.5, 0.5)
+	imgui.GetStyle().SelectableTextAlign = imgui.ImVec2(0, 0)
 	
 	--==[ COLORS ]==--
 	imgui.GetStyle().Colors[imgui.Col.Text]                   = imgui.ImVec4(1.00, 1.00, 1.00, 1.00)
